# server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: NLM Service API
  description: API specification for nlm-service
  contact:
    email: support@nlmatics.com
  version: 2.2.13
servers:
  - url: http://localhost:5000/api
    description: NLMATICS Service API Endpoint
...
paths:
...
  /user:
#    post:
#      tags:
#        - user
#      summary: Creates a new user
#      operationId: create_user
#      requestBody:
#        content:
#          application/json:
#            schema:
#              $ref: '#/components/schemas/User'
#      responses:
#        "200":
#          description: Returns the unique id assigned to the new user
#          content:
#            application/json:
#              schema:
#                $ref: '#/components/schemas/IdWithMessage'
#      x-openapi-router-controller: server.controllers.user_controller
    get:
      tags:
        - user
      summary: Fetches user info by emailid
      operationId: get_user_by_email
      parameters:
        - name: emailId
          in: query
          description: user emailid
          required: true
          explode: false
          schema:
            type: string
      responses:
        200:
          description: "Returns the user info"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      x-openapi-router-controller: server.controllers.user_controller
...
  /user/{userId}:
    get:
      tags:
        - user
      summary: Returns information about the user with id
      operationId: get_user_info_by_id
      parameters:
        - name: userId
          in: path
          description: user id
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: Return information about the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      x-openapi-router-controller: server.controllers.user_controller

#  /user/modify/{userId}:
#    post:
#      tags:
#        - user
#      summary: Update an existing user with id
#      operationId: update_user_by_id
#      parameters:
#        - name: userId
#          in: path
#          required: true
#          style: simple
#          explode: false
#          schema:
#            type: string
#      requestBody:
#        content:
#          application/json:
#            schema:
#              $ref: '#/components/schemas/User'
#      responses:
#        "200":
#          description: User id of the updated user
#          content:
#            application/json:
#              schema:
#                $ref: '#/components/schemas/IdWithMessage'
#      x-openapi-router-controller: server.controllers.user_controller
#
#  /user/delete/{userId}:
#    post:
#      tags:
#        - user
#      summary: Delete user with given id
#      operationId: delete_user_by_id
#      parameters:
#        - name: userId
#          in: path
#          required: true
#          style: simple
#          explode: false
#          schema:
#            type: string
#      responses:
#        "200":
#          description: user id of the deleted user
#          content:
#            application/json:
#              schema:
#                $ref: '#/components/schemas/IdWithMessage'
#      x-openapi-router-controller: server.controllers.user_controller
...
security:
  - jwt: ['admin']
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        emailId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        active:
          type: boolean
        isAdmin:
          type: boolean
        createdOn:
          type: string
          format: date-time
        hasDeveloperAccount:
          type: boolean
          default: false
        developerAppId:
          type: string
        expiry_time:
          type: string
          format: date-time
        subscription_plan:
          type: string
        accessType:
          type: string
        lastLogin:
          type: string
          format: date-time
        hasNotifications:
          type: boolean
          default: false
        workspaceNotificationSettings:
          type: array
          items:
            type: object
        subscribedWorkspaces:
          type: array
          items:
            type: string
        includedFeatures:
          type: array
          items:
            type: string
        stripeConf:
          type: object
      example:
        firstName: firstName
        lastName: lastName
        active: true
        emailId: emailId
        id: id
        createdOn: 2000-01-23T04:56:07.000+00:00
        hasDeveloperAccount: false
        expiry_time: 2031-08-30 21:41:55 EST-0500
        subscription_plan: BASIC
        lastLogin: 2021-08-30 21:41:55 EST-0500
    IdWithMessage:
      type: object
      properties:
        id:
          type: string
        message:
          type: string
      example:
        id: id
        message: message
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: server.auth.authenticate